Data - Add Data Settings - 9405663157 (#91)
* Data - Settings - Initial Commit

* Data - Update Global Imports

* Data - Update Imports

* Misc - Clean Up Domain Imports and Tests
