+++ tiferet/data/__init__.py
@@ -1,5 +1,4 @@
-from typing import List, Dict
+# *** imports
 
-from .error import *
-from .feature import *
-from .app import *
+# ** app
+from .settings import *
