+++ tiferet/data/container.py
@@ -1,11 +1,8 @@
 # *** imports
 
-# ** core
-from typing import Any, Dict
-
 # ** app
 from ..configs import *
-from ..domain import DataObject
+from ..data import DataObject
 from ..domain.container import ContainerAttribute, ContainerDependency
